[BinFile] Handle zero-virtual-size PE sections
Some PE section header indicate that the corresponding section has zero
virtual size even if there is really a data. For this reason, we cannot
simply rely on the GetContainingSectionIndex[^1] function as it does not
match a section where its header indicates zero virtual size. Thus, this
commit adds a new function that falls back on the
GetContainingSectionIndex by considering the raw section size
(SizeOfRawData).

[^1] https://docs.microsoft.com/en-us/dotnet/api/system.reflection.portableexecutable.peheaders.getcontainingsectionindex?view=netcore-2.2#System_Reflection_PortableExecutable_PEHeaders_GetContainingSectionIndex_System_Int32_