[FrontEnd] Add GetNextInstrAddrs method to instr
We should not rely on `IsCondBranch()` to determine the target address
of a conditional branch instruction, because in some architectures,
conditional branches can have three or more targets, or they may not
have a fall-through target. Therefore, we should create GetNextAddrs
method for each architecture, so that we can easily get potential target
addresses of an instruction regardless of the instruction kind.