[MiddleEnd] Fix intra leader insertion logic
When we construct a list of intra-leaders, we do not sort them for the
performance reason. However, we were simply prepending a new
intra-leader to the list when we encounter an instruction like `JMP PC`.
This commit fixes the bug making our leader scanning algorithm to
correctly construct a sorted list.