[MiddleEnd] Update jumptable cancellation logic
When an identified jump table entry points to an invalid address, we
should be able to either rollback or cancel the process. In particular,
when the first jump table entry is wrong, then we consider the
corresponding indirect branch to be a falsely identified case (e.g., a
function pointer) and cancel the jump table recovery process.