[MiddleEnd] Properly unregister jump tables
To handle racing conditions, modified JumpTableNotes.Unregister function
to take a function address as input. This is because two different
functions can share the same (overlapped) indirect branch in some rare
cases, and we may hit racing conditions when the one of the functions
tries to update JumpTableNotes while another one tries to unregister
the same note.