[MiddleEnd] Use trampoline's address for PLT
Make it more intuitive and fix GUI later.

We used `TableAddress` (i.e., GOT address) as a key for an external
builder in order to consider a GOT entry as an external function. But it
is more natural and intuitive to consider a PLT entry (not a GOT entry)
as an external function.