[ELF] Improve PLT entry size calculation
PLT sections in some binaries may contain extra data, which can cause
our logic for computing the size of PLT entries to fail. This is
particularly true for old PPC32 binaries. This patch directly uses
relocation information to determine the size of PLT entries, and falls
back to the old logic if there are less than two PLT relocations. With
this change, we no longer need to special case old PPC32 binaries.